refactor(naming): rename banned-word files + EnhancedAnalyticsGrid component (#10746)
Remove banned Enhanced/Unified words from file names and the analytics
Vue component name; two historical migration scripts kept (targets are
canonically named).

- EnhancedAnalyticsGrid.{vue,stories.ts} -> AnalyticsGrid.* (+ barrel export,
  story title/refs, root CSS class)
- test_unified_progress_tracker -> test_progress_tracker
- test_credential_store_unified_read -> test_credential_store_read (migrations + connectors)
- test_credential_store_unified_write -> test_credential_store_write
- test_unified_secrets_api -> test_secrets_api
- test_unified_secrets_service -> test_secrets_service
