+++ autobot-backend/tests/unit/knowledge/connectors/test_credential_store_write.py
